+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## [Unreleased]
+### Changed
+- Deprecate `Knock::Authenticable` in favor of `Knock::Authenticatable`
+
 ## [1.4.2] - 2016-01-29
 ### Fixed
 - Allow use of any or no prefix in authorization header.